BE/Refactor : Math.clamp 이용할 수 있도록 수정

+++ backend/src/main/java/com/naengsam/quick/domain/matching/policy/assignment/MatchingAssignmentProblemAssembler.java
@@ -126,6 +126,6 @@ public class MatchingAssignmentProblemAssembler {
         }
 
         long quota = ((long) dreamiCount + orderCount - 1) / orderCount;
-        return (int) Math.clamp(quota, 1, 5);
+        return Math.clamp(quota, 1, 5);
     }
 }
